backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
PyJWT[crypto]==2.10.1
httpx==0.28.1
SQLAlchemy==2.0.38
alembic==1.14.1
psycopg[binary]==3.2.9
python-multipart==0.0.12
minio==7.2.15
